Restructure monitored projects
So far we had two pages related to external projects
(extension-support and release dates).
However, this topic-based structure is not optimal
to provide a quick status overview of a single project.

Let's restructure the way we collect monitoring information
of external projects by concentrating all information relevant
for a project on a single place. To not end up with a single
endless list, let's group the items into four sets.

Signed-off-by: Christoph Muellner <christophm30@gmail.com>